feat: make jsluice optional behind build tag
